+++ tests/util.rs
@@ -4,7 +4,7 @@ use std::fs::{copy, create_dir};
 use std::io::Result;
 
 pub fn run_command(args: Vec<&str>, config_file: Option<&str>) -> Result<()> {
-    let mut args = args; // Borrow checker bug (I think)
+    let mut args = args;
     let running = format!("./tests/configs/running/{}.json", rand::random::<u16>());
     if let Some(config_file) = config_file {
         let _ = create_dir("./tests/configs/running");
@@ -14,10 +14,6 @@ pub fn run_command(args: Vec<&str>, config_file: Option<&str>) -> Result<()> {
 
     let mut command = Command::new(env!("CARGO_BIN_EXE_ferium"));
     let mut arguments = Vec::new();
-    if let Some(token) = option_env!("GITHUB_TOKEN") {
-        arguments.push("--github-token");
-        arguments.push(token);
-    }
     arguments.push("--config-file");
     arguments.push(&running);
     arguments.append(&mut args);